+++ Src/Library/Endpoint/Summary/ProducesResponseTypeMetadata.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Http.Metadata;
+
+namespace FastEndpoints;
+
+internal class ProducesResponseTypeMetadata : IProducesResponseTypeMetadata
+{
+    public Type? Type { get; set; }
+
+    public int StatusCode { get; set; }
+
+    public IEnumerable<string> ContentTypes { get; set; }
+
+    public object? Example { get; set; }
+}
